+++ Delta.CertXplorer/src/Delta.CapiNet/Asn1/asn1Types/Asn1NumericString.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Delta.CapiNet.Asn1
+﻿namespace Delta.CapiNet.Asn1
 {
     public class Asn1NumericString : Asn1Object
     {