+++ Delta.CertXplorer/src/Delta.CapiNet/Asn1/asn1Types/Asn1OctetString.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Delta.CapiNet.Asn1
+﻿namespace Delta.CapiNet.Asn1
 {
     // According to PKCS, octet strings may be structures...
     // This is why we try to find children in it.