+++ Delta.CertXplorer/src/Delta.CapiNet/Asn1/asn1Types/Asn1PrintableString.cs
@@ -1,21 +1,15 @@
-﻿using System;
+﻿using System.Text;
 
 namespace Delta.CapiNet.Asn1
 {
-    internal class Asn1PrintableString : Asn1Utf8String
+    internal class Asn1PrintableString : BaseAsn1String
     {
+        private static readonly Encoding windows1252 = Encoding.GetEncoding(1252);
+
         public Asn1PrintableString(Asn1Document document, TaggedObject content, Asn1Object parentObject)
             : base(document, content, parentObject) { }
 
-        /// <summary>
-        /// Returns a <see cref="System.String"/> that represents this instance.
-        /// </summary>
-        /// <returns>
-        /// A <see cref="System.String"/> that represents this instance.
-        /// </returns>
-        public override string ToString()
-        {
-            return string.Format("PrintableString: {0}", base.Value);
-        }
+        protected override string FriendlyName => "PrintableString";
+        protected override Encoding Encoding => windows1252;
     }
 }