+++ Delta.CertXplorer/src/Delta.CapiNet/Asn1/asn1Types/Asn1Utf8String.cs
@@ -1,33 +1,10 @@
-﻿using System;
-using System.Text;
-
-namespace Delta.CapiNet.Asn1
+﻿namespace Delta.CapiNet.Asn1
 {
-    public class Asn1Utf8String : Asn1Object
+    public sealed class Asn1Utf8String : BaseAsn1String
     {
-        private string utf8 = string.Empty;
-
         public Asn1Utf8String(Asn1Document document, TaggedObject content, Asn1Object parentObject)
-            : base(document, content, parentObject)
-        {
-            utf8 = Encoding.UTF8.GetString(content.Workload); 
-        }
+            : base(document, content, parentObject) { }
 
-        /// <summary>
-        /// Gets this instance's value.
-        /// </summary>
-        /// <value>The value.</value>
-        public string Value { get { return utf8; } }
-
-        /// <summary>
-        /// Returns a <see cref="System.String"/> that represents this instance.
-        /// </summary>
-        /// <returns>
-        /// A <see cref="System.String"/> that represents this instance.
-        /// </returns>
-        public override string ToString()
-        {
-            return string.Format("Utf8String: {0}", utf8);
-        }
+        protected override string FriendlyName => "Utf8String";
     }
 }